PruneMsgsUntil unit test
